follow/unfollow tags, global feed, personal feed

+++ server/controllers/post.controller.ts
@@ -3,20 +3,6 @@ import PostModel from "../models/post.model";
 import { Created, NoContent, Ok } from "../lib/utils/express.utils";
 import { NotFoundError } from "../lib/api.errors";
 import UserModel from "../models/user.model";
-import { CreatePostBodySchema } from "../lib/schemas/post/create-post.schema";
-import { DeletePostParamsSchema } from "../lib/schemas/post/delete-post.schema";
-import {
-  GetPostParamsSchema,
-  GetPostListParamsSchema,
-} from "../lib/schemas/post/get-post.schema";
-import {
-  UpdatePostParamsSchema,
-  UpdatePostBodySchema,
-} from "../lib/schemas/post/update-post.schema";
-import {
-  FavoritePostParamsSchema,
-  UnfavortiePostParamsSchema,
-} from "../lib/schemas/post/favorite-post.schema";
 import { UserProfileDTO } from "../lib/types/user.types";
 import TagModel from "../models/tag.model";
 import utils from "../lib/utils/utils";
@@ -27,13 +13,31 @@ import {
   mapToPostDTO,
 } from "../lib/types/post.types";
 import { PagedResponse } from "../lib/types";
+import mongoose from "mongoose";
+import {
+  CreatePostBodySchema,
+  DeletePostParamsSchema,
+  FavoritePostParamsSchema,
+  GetPostListParamsSchema,
+  GetPostParamsSchema,
+  UnfavortiePostParamsSchema,
+  UpdatePostBodySchema,
+  UpdatePostParamsSchema,
+} from "../lib/schemas/post";
 
+/**
+ * Creates a new post.
+ *
+ * @param req - The request object.
+ * @param res - The response object.
+ * @returns The created post.
+ * @throws {NotFoundError} if the user is not found.
+ */
 async function create(req: Request, res: Response) {
   const reqBody = CreatePostBodySchema.parse(req.body);
 
   const userId = req.user?.id;
 
-  // any first user
   const user = await UserModel.findById(userId);
 
   if (!user) {
@@ -46,6 +50,8 @@ async function create(req: Request, res: Response) {
     readingTime: utils.readingTime(reqBody.body),
     ...reqBody,
   });
+
+  await user.addPost(createdPost._id);
 
   await TagModel.updateMany(
     { _id: { $in: createdPost.tags } },
@@ -79,6 +85,15 @@ async function getBySlug(req: Request, res: Response) {
   return Ok(res, mappedPost);
 }
 
+/**
+ * Retrieves a list of posts based on the provided query parameters.
+ *
+ * @param req - The request object.
+ * @param res - The response object.
+ * @returns A response containing the list of posts and pagination information.
+ * @throws {Error} If both tag and following are specified.
+ * @throws {NotFoundError} If the specified tag or user is not found.
+ */
 async function getList(req: Request, res: Response) {
   const {
     page = 1,
@@ -86,22 +101,42 @@ async function getList(req: Request, res: Response) {
     query,
     orderBy,
     tag: tagSlug,
+    feed,
   } = GetPostListParamsSchema.parse(req.query);
 
-  let tag = undefined;
+  const isUserFeed = feed === "user";
+
+  if (tagSlug && isUserFeed) {
+    throw new Error("Only one of tag or following can be specified");
+  }
+
+  let tags: mongoose.Types.ObjectId[] = [];
+
   if (tagSlug) {
-    tag = await TagModel.findOne({ slug: tagSlug });
+    const tag = await TagModel.findOne({ slug: tagSlug });
     if (!tag) {
       throw new NotFoundError(`Tag with slug ${tagSlug} not found`);
     }
+
+    tags = [tag._id];
+  }
+
+  if (isUserFeed) {
+    const userId = req.user?.id;
+    const user = await UserModel.findById(userId);
+    if (!user) {
+      throw new NotFoundError(`User with id ${userId} not found`);
+    }
+
+    tags = user.followingTags;
   }
 
   const searchOptions = Object.assign(
     {},
     query && { title: { $regex: query, $options: "i" } },
-    tag && {
+    (isUserFeed || tags.length > 0) && {
       tags: {
-        $in: [tag._id],
+        $in: tags,
       },
     }
   );
@@ -155,11 +190,19 @@ async function getList(req: Request, res: Response) {
   return Ok(res, pagedResponse);
 }
 
+/**
+ * Updates a post by its identifier.
+ *
+ * @param req - The request object.
+ * @param res - The response object.
+ * @returns The updated post.
+ * @throws {NotFoundError} if the post with the given identifier is not found.
+ */
 async function updateById(req: Request, res: Response) {
   const reqParams = UpdatePostParamsSchema.parse(req.params);
   const reqBody = UpdatePostBodySchema.parse(req.body);
 
-  const postToUpdate = await PostModel.findOne({ _id: reqParams.identifier });
+  const postToUpdate = await PostModel.findById(reqParams.identifier);
 
   if (!postToUpdate) {
     throw new NotFoundError(`Post with id ${reqParams.identifier} not found`);
@@ -188,13 +231,34 @@ async function updateById(req: Request, res: Response) {
   return Ok(res, postToUpdate);
 }
 
+/**
+ * Deletes a post by its identifier.
+ *
+ * @param req - The request object.
+ * @param res - The response object.
+ * @returns A response with no content.
+ * @throws {NotFoundError} if the user or post is not found.
+ */
 async function deleteById(req: Request, res: Response) {
   const reqParams = DeletePostParamsSchema.parse(req.params);
-  const postToDelete = await PostModel.findOne({ _id: reqParams.identifier });
+
+  const userId = req.user!.id;
+
+  const user = await UserModel.findById(userId);
+
+  if (!user) {
+    throw new NotFoundError(`User with id ${userId} not found`);
+  }
+
+  const postToDelete = await PostModel.findById(reqParams.identifier);
+
   if (!postToDelete) {
     throw new NotFoundError(`Post with id ${reqParams.identifier} not found`);
   }
+
   await postToDelete.deleteOne();
+
+  await user.removePost(postToDelete._id);
 
   await TagModel.updateMany(
     { _id: { $in: postToDelete.tags } },
@@ -204,24 +268,52 @@ async function deleteById(req: Request, res: Response) {
   return NoContent(res);
 }
 
+/**
+ * Handles the request to favorite a post.
+ *
+ * @param req - The request object.
+ * @param res - The response object.
+ * @returns A response indicating the success of the operation.
+ * @throws {NotFoundError} if the user or post is not found.
+ */
 async function favorite(req: Request, res: Response) {
   const reqParams = FavoritePostParamsSchema.parse(req.params);
   const userId = req.user!.id;
 
+  const user = await UserModel.findById(userId);
+  if (!user) {
+    throw new NotFoundError(`User with id ${userId} not found`);
+  }
+
   const foundPost = await PostModel.findById(reqParams.identifier);
 
   if (!foundPost) {
     throw new NotFoundError(`Post with id ${reqParams.identifier} not found`);
   }
 
-  await foundPost.addFavorite(userId);
+  await foundPost.addToPostFavorites(user._id);
+  await user.addFavoritePost(foundPost._id);
 
   return NoContent(res);
 }
 
+/**
+ * Unfavorites a post.
+ *
+ * @param req - The request object.
+ * @param res - The response object.
+ * @returns The response with no content.
+ * @throws {NotFoundError} if the user or post is not found.
+ */
 async function unfavorite(req: Request, res: Response) {
   const reqParams = UnfavortiePostParamsSchema.parse(req.params);
   const userId = req.user!.id;
+
+  const user = await UserModel.findById(userId);
+
+  if (!user) {
+    throw new NotFoundError(`User with id ${userId} not found`);
+  }
 
   const foundPost = await PostModel.findById(reqParams.identifier);
 
@@ -229,7 +321,8 @@ async function unfavorite(req: Request, res: Response) {
     throw new NotFoundError(`Post with id ${reqParams.identifier} not found`);
   }
 
-  await foundPost.removeFavorite(userId);
+  await foundPost.removeFromPostFavorites(user._id);
+  await user.removeFavoritePost(foundPost._id);
 
   return NoContent(res);
 }
